+++ components/icons/SecurityHotspotIcon.tsx
@@ -18,17 +18,15 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 import * as React from 'react';
-import { IconProps, ThemedIcon } from './Icon';
+import Icon, { IconProps } from './Icon';
 
-export default function SecurityHotspotIcon({ className, fill, size }: IconProps) {
+export default function SecurityHotspotIcon({ className, fill = 'currentColor', size }: IconProps) {
   return (
-    <ThemedIcon className={className} size={size}>
-      {({ theme }) => (
-        <path
-          d="M14.08 3.23a1 1 0 00-.67-.77L8.16 1a1.06 1.06 0 00-.5 0L2.41 2.46a.94.94 0 00-.67.77c-.08.57-.74 5.63 1.14 8.31A9 9 0 007.68 15a.85.85 0 00.23 0 .78.78 0 00.22 0 8.93 8.93 0 004.81-3.46c1.85-2.68 1.21-7.74 1.14-8.31zM12.21 8a6.15 6.15 0 01-.86 2.42A7.92 7.92 0 018 13V8zM8 3v5H3.59a24.29 24.29 0 010-3.82z"
-          style={{ fill: fill || theme.colors.baseFontColor }}
-        />
-      )}
-    </ThemedIcon>
+    <Icon className={className} size={size}>
+      <path
+        d="M14.08 3.23a1 1 0 00-.67-.77L8.16 1a1.06 1.06 0 00-.5 0L2.41 2.46a.94.94 0 00-.67.77c-.08.57-.74 5.63 1.14 8.31A9 9 0 007.68 15a.85.85 0 00.23 0 .78.78 0 00.22 0 8.93 8.93 0 004.81-3.46c1.85-2.68 1.21-7.74 1.14-8.31zM12.21 8a6.15 6.15 0 01-.86 2.42A7.92 7.92 0 018 13V8zM8 3v5H3.59a24.29 24.29 0 010-3.82z"
+        style={{ fill }}
+      />
+    </Icon>
   );
 }